Implemented runtime configuration
EML drivers are now configurable through an optional config file.

libconfuse is used for now to parse the configs.
This may change in the future so we can avoid one current significant
pitfall: configuration files can't define config sections for drivers
that are not compiled in, or unknown in the current version. This
precludes use of the same configuration file in different setups.

Please refer to doc/configuration.md for details.