Added initial Schleifenbauer PDU support
This commit introduces the sb_pdu driver with partial support for
Schleifenbauer power distribution units. It uses the socket API
implemented in: http://schleifenbauer.eu/downloads/API/

PDUs are queried over TCP/IP. They are not autodetected and should be
declared in the sb_pdu section of the configuration file:

	sb_pdu {
	  device ourpdu {
	    host = 10.0.0.3
	    port = 7783
	  }
	}

The driver depends on libcrypto for the RC4 encryption. Although perhaps
disproportionate, this dependency may be useful in the future if support
is added for any other devices using encryption.

Current limitations and known issues:
- only one PDU module per configured device is (properly) supported
- socket operations are blocking for now
- the implemented measurement caching introduces avoidable inaccuracy
- existing pdu-level energy counters are not used